fixed time display

+++ app/src/main/java/com/john/gotouchgrass/model/GrassMoment.kt
@@ -7,7 +7,7 @@ import java.util.*
  */
 data class GrassMoment(
     // The time a user spent outside.
-    val timeSpentOutside: Int,
+    val timeSpentOutside: Double,
     // The user's written description of their time outside.
     val description: String,
     // The rating a user gave their time outside (OKAY, BAD, GREAT).
